pyproject.toml
[build-system]
requires = ["setuptools"]


[project]
name = "rte"
version = "0.1.0"

dependencies = [
    "grpcio",
    "grpcio-tools",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pylint",
    "mypy",
]
test = [
    "pytest",
]

[tool.ruff]
line-length = 100

[tool.pylint]
disable = [
    "C0114",  # Missing module docstring
    "C0115",  # Missing class docstring
    "C0116",  # Missing function or method docstring
    "E0611",  # No name '...' in module '...'
]

[tool.pytest.ini_options]
filterwarnings = [
    "error", # treat all warnings as errors
    "ignore:.*custom tp_new.*in Python 3.14.*:DeprecationWarning",
]
